requirements.txt
#
# Python library requirements for netsim-tools
#
Jinja2 >= 2.11.0
PyYAML >= 5.0.0
netaddr >= 0.8.0
python-box ~= 7.0
importlib_resources
typing-extensions>=4.3.0
filelock >= 3.10
packaging
rich
